move api tests too
